+++ src/components/chatroom/RoomHeader.tsx
@@ -1,0 +1,5 @@
+const RoomHeader = () => {
+    return <></>;
+};
+
+export default RoomHeader;
