+++ src/pages/Chatroom.tsx
@@ -1,9 +1,16 @@
 import ChatItem from '../components/chatroom/ChatItem';
 import { ChatType } from '../Interface';
-import useApi from '../hooks/useApi';
+import * as api from '../api';
+import { useState } from 'react';
+
 const ChatRoom = ({ roomId }: { roomId: number }) => {
-    const { getChats } = useApi();
-    const chats = getChats(roomId);
+    const { GetChats, GetUserInfoListByRoomId } = api;
+    const chats = GetChats(roomId);
+    const userList = GetUserInfoListByRoomId(roomId);
+    console.log(userList);
+    const [me, setMe] = useState(userList[0]);
+    const toggleMe = () => {};
+
     return (
         <>
             {chats.map((chat: ChatType) => (
